+++ src/AV_Spex/utils/dir_setup.py
@@ -1,0 +1,140 @@
+import os
+import sys
+import shutil
+from ..utils.log_setup import logger
+
+def validate_input_paths(input_paths, is_file_mode):
+    source_directories = []
+    for input_path in input_paths:
+        try:
+            if is_file_mode and not os.path.isfile(input_path):
+                raise ValueError(f"Error: {input_path} is not a valid file.")
+            elif not is_file_mode and not os.path.isdir(input_path):
+                raise ValueError(f"Error: {input_path} is not a valid directory.")
+            
+            directory = os.path.dirname(input_path) if is_file_mode else input_path
+            source_directories.append(directory)
+            logger.info(f'Input directory found: {directory}\n')
+        except ValueError as e:
+            logger.critical(str(e))
+            sys.exit(1)
+    return source_directories
+
+
+def move_vrec_files(directory, video_id):
+    vrecord_files_found = False
+
+    # Create the target directory path
+    vrecord_directory = os.path.join(directory, f'{video_id}_vrecord_metadata')
+
+    # Check if the vrecord directory already exists and contains the expected files
+    if os.path.exists(vrecord_directory):
+        expected_files = [
+            '_QC_output_graphs.jpeg',
+            '_vrecord_input.log',
+            '_capture_options.log',
+            '.mkv.qctools.mkv',
+            '.framemd5'
+        ]
+
+        # Check if at least one expected file is in the vrecord directory
+        if any(filename.endswith(ext) for ext in expected_files for filename in os.listdir(vrecord_directory)):
+            logger.debug(f"Existing vrecord files found in {os.path.basename(directory)}/{os.path.basename(vrecord_directory)}\n")
+            return
+
+    # Iterate through files in the directory
+    for filename in os.listdir(directory):
+        file_path = os.path.join(directory, filename)
+
+        # Check if the file matches the naming convention
+        if (
+            os.path.isfile(file_path)
+            and filename.endswith(('_QC_output_graphs.jpeg', '_vrecord_input.log', '_capture_options.log', '.mkv.qctools.mkv', '.framemd5'))
+        ):
+            # Create the target directory if it doesn't exist
+            vrecord_directory = os.path.join(directory, f'{video_id}_vrecord_metadata')
+            os.makedirs(vrecord_directory, exist_ok=True)
+            # Move the file to the target directory
+            new_path = os.path.join(vrecord_directory, filename)
+            shutil.move(file_path, new_path)
+            # logger.debug(f'Moved vrecord file: {filename} to directory: {os.path.basename(vrecord_directory)}')
+            vrecord_files_found = True
+
+    # Check if any matching files were found to create the directory
+    if vrecord_files_found:
+        logger.debug(f"Files generated by vrecord found. '{video_id}_vrecord_metadata' directory created and files moved.\n")
+    else:
+        logger.debug("No vrecord files found.\n")
+
+
+def find_mkv(source_directory):
+    # Create empty list to store any found mkv files
+    found_mkvs = []
+    for filename in os.listdir(source_directory):
+        if filename.lower().endswith('.mkv'):
+            if 'qctools' not in filename.lower():
+                found_mkvs.append(filename)
+    # check if found_mkvs is more than one
+    if found_mkvs:
+        if len(found_mkvs) == 1:
+            video_path = os.path.join(source_directory, found_mkvs[0])
+            logger.info(f'Input video file found in {source_directory}: {video_path}\n')
+        else:
+            logger.critical(f'More than 1 mkv found in {source_directory}: {found_mkvs}\n')
+            return None
+    else:
+        logger.critical(f"Error: No mkv video file found in the directory: {source_directory}\n")
+        return None
+
+    return video_path
+
+
+def check_directory(source_directory, video_id):
+    """
+    Checks whether the base name of a directory matches the given video_id.
+    
+    Args:
+        source_directory (str): The path to the directory.
+        video_id (str): The expected video ID to match.
+
+    Returns:
+        bool: True if the directory name matches the video_id, otherwise False.
+    """
+    directory_name = os.path.basename(source_directory)
+    if directory_name.startswith(video_id):
+        logger.info(f'Directory name "{directory_name}" correctly matches video file name "{video_id}".\n')
+        return True
+    else:
+        logger.critical(f'Directory name "{directory_name}" does not correctly match the expected "{video_id}".\n')
+        return False
+
+
+def make_qc_output_dir(source_directory, video_id):
+    '''
+    Creates output directory for metadata files
+    '''
+
+    destination_directory = os.path.join(source_directory, f'{video_id}_qc_metadata')
+
+    if not os.path.exists(destination_directory):
+        os.makedirs(destination_directory)
+
+    logger.debug(f'Metadata files will be written to {destination_directory}\n')
+
+    return destination_directory
+
+
+def make_report_dir(source_directory, video_id):
+    '''
+    Creates output directory for metadata files
+    '''
+
+    report_directory = os.path.join(source_directory, f'{video_id}_report_csvs')
+
+    if os.path.exists(report_directory):
+        shutil.rmtree(report_directory)
+    os.makedirs(report_directory)
+
+    logger.debug(f'Report files will be written to {report_directory}\n')
+
+    return report_directory
